move into mpi utils
