+++ fv3core/utils/gt4py_utils.py
@@ -10,6 +10,7 @@ import numpy as np
 from gt4py import gtscript
 
 from fv3core.utils.mpi import MPI
+from fv3core.utils import mpi
 from fv3core.utils.typing import DTypes, Field, float_type, int_type
 
 from . import global_config
@@ -53,60 +54,8 @@ tracer_variables = [
 logger = logging.getLogger("fv3ser")
 
 
-def get_size() -> int:
-    """Get the size of the MPI COMM_WORLD communicator if it is defined.
-
-    Returns:
-        int: size
-    """
-    return MPI.COMM_WORLD.Get_size() if MPI else 0
-
-
-def get_rank() -> int:
-    """Get the current rank in the MPI COMM_WORLD communicator if it is defined.
-
-    Returns:
-        int: rank
-    """
-    return MPI.COMM_WORLD.Get_rank() if get_size() > 1 else -1
-
-
-def bcast(data) -> None:
-    """A wrapper for a broadcast in the MPI COMM_WORLD communicator from rank 0.
-
-    Args:
-        data: The data to broadcast
-    """
-    return MPI.COMM_WORLD.bcast(data, root=0)
-
-
-def send(dest: int, tag: int = 0) -> None:
-    """A wrapper for a dummy send in the MPI COMM_WORLD communicator
-    used for blocking other threads from advancing.
-
-    Args:
-        dest: The destination rank
-        tag: The tag that the MPI send should have
-    """
-    MPI.COMM_WORLD.send(1, dest=dest, tag=tag)
-
-
-def recv(source: int, tag: int = 0) -> Any:
-    """A wrapper for a dummy receive in the MPI COMM_WORLD communicator used for
-    blocking other threads from advancing.
-
-    Args:
-        source: The source rank
-        tag: The tag that the MPI send will have
-    Returns:
-        Forwadrd the recv return value
-    """
-    buff = bytearray(1 << 10)
-    return MPI.COMM_WORLD.recv(buff, source=source, tag=tag)
-
-
 # 1 indexing to 0 and halos: -2, -1, 0 --> 0, 1,2
-if get_size() > 1:
+if mpi.get_size() > 1:
     gt.config.cache_settings["dir_name"] = ".gt_cache_{:0>6d}".format(
         MPI.COMM_WORLD.Get_rank()
     )
